Functional parallelized backend, needs testing.
Added globallock as alternative to advanced lock.
Further parallelized sharerlock and sharerWindow to allow
more concurrent writes to the same node.
Simple padding of globalDataWindow to stop what seems to be
false sharing? Should the same be applied to sharerWindow?
Borrowed implementation of stripe allocation to allow for
better load balancing across nodes.
TODO: Does not work for IMPI <2019.x