Improved ArgoDSM write buffer
This commit adds a dedicated ArgoDSM write buffer that now acts
as the base for deciding whether or not to write back pages
instead of iterating over the entire cache checking for dirty
pages. It also adds a foundation for selective coherence correctly
removing downgraded pages from the write buffer.