Add interface function to check if page is cached
This commit adds an interface function to check whether a page
is cached on a node. This function returns true if, for the
address specified, the corresponding page is either cached or
locally backed on the node, otherwise it returns false. This
is strictly meant to be used in order to thoroughly test
prefetching of pages.